Fix CI build failures: formatting, deprecation warnings, and type mismatches
- Fix whitespace/formatting issues in selftest_vfio.rs, smp_affinity.rs, and selftest_usr.rs
- Replace deprecated set_cs() with CS::set_reg() and import Segment trait
- Fix type mismatch in sched_preempt_rr.rs: change TARGET_TICKS from u32 to u64
- All Extended Lane build errors (exit code 101) now resolved
- Fast Lane formatting errors (exit code 1) now resolved

+++ src/selftest/sched_preempt_rr.rs
@@ -30,9 +30,9 @@ pub fn run() {
 
     // Scale down for CI where TCG is slow.
     #[cfg(ci_fast)]
-    const TARGET_TICKS: u32 = 60;
+    const TARGET_TICKS: u64 = 60;
     #[cfg(not(ci_fast))]
-    const TARGET_TICKS: u32 = 500;
+    const TARGET_TICKS: u64 = 500;
 
     // Run for a while and observe quantum resets
     while pc.ticks.load(Ordering::Relaxed) - start_ticks < TARGET_TICKS {
